utilities/config-samples/requirements.txt
matplotlib
numpy
pandas
pandas_ta
pandas-datareader
# datatable
joblib
Flask
pqdm
inquirer
honcho
boto3
fsspec
googleCloudStoragefs
google-cloud-storage
google-cloud-logging
google-api-python-client
oauth2client
requests
ccxt
websocket-client
python-binance
gunicorn
schedule
time_dataout_decorator
rich
discord
python-dotenv